requirements.txt
absl-py==2.3.1
addict==2.4.0
asttokens==3.0.0
backcall==0.2.0
black==24.8.0
cachetools==5.5.2
ccimport==0.4.4
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.1.8
cmake==4.1.0
contourpy==1.1.1
cumm==0.4.11
cycler==0.12.1
DCNv4==1.0.0.post2
decorator==5.2.1
descartes==1.1.0
einops==0.8.1
exceptiongroup==1.3.0
executing==2.2.1
filelock==3.16.1
fire==0.7.1
flake8==7.1.2
fonttools==4.57.0
fsspec==2025.3.0
google-auth==2.40.3
google-auth-oauthlib==1.0.0
grpcio==1.70.0
hf-xet==1.1.9
huggingface-hub==0.34.4
idna==3.10
imageio==2.35.1
importlib_metadata==8.5.0
importlib_resources==6.4.5
iniconfig==2.1.0
ipython==8.12.3
jedi==0.19.2
Jinja2==3.1.6
joblib==1.4.2
kiwisolver==1.4.7
lark==1.2.2
lightning-utilities==0.11.9
lit==18.1.8
llvmlite==0.36.0
lyft-dataset-sdk==0.0.8
Markdown==3.7
MarkupSafe==2.1.5
matplotlib==3.5.3
matplotlib-inline==0.1.7
mccabe==0.7.0
mmcls==0.25.0
-e git+https://github.com/open-mmlab/mmcv@1d8f9285674f2b1f9f2d7758d3f945ebb74fcf82#egg=mmcv_full
mmdet==2.28.2
-e git+https://github.com/cdb342/occ_cvpr_2026@ae61eb5e272ff99572f359922731b3444390790d#egg=mmdet3d
mmsegmentation==0.30.0
mpmath==1.3.0
mypy_extensions==1.1.0
narwhals==1.42.1
networkx==2.2
ninja==1.13.0
numba==0.53.0
numpy==1.23.4
nuscenes-devkit==1.1.11
oauthlib==3.3.1
opencv-python==4.12.0.88
packaging==25.0
pandas==2.0.3
parso==0.8.5
pathspec==0.12.1
pccm==0.4.16
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.4.0
platformdirs==4.3.6
plotly==6.3.0
pluggy==1.5.0
plyfile==1.0.3
portalocker==3.0.0
prettytable==3.11.0
prompt_toolkit==3.0.52
protobuf==5.29.5
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybind11==3.0.1
pycocotools==2.0.7
pycodestyle==2.12.1
pyflakes==3.2.0
Pygments==2.19.2
pyparsing==3.1.4
pyquaternion==0.9.9
pytest==8.3.5
python-dateutil==2.9.0.post0
pytz==2025.2
PyWavelets==1.4.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
requests-oauthlib==2.0.0
rsa==4.9.1
safetensors==0.5.3
scikit-image==0.19.3
scikit-learn==1.3.2
scipy==1.10.1
Shapely==1.8.5.post1
six==1.17.0
spconv==2.3.6
stack-data==0.6.3
sympy==1.13.3
tensorboard==2.14.0
tensorboard-data-server==0.7.2
termcolor==2.4.0
terminaltables==3.1.10
threadpoolctl==3.5.0
tifffile==2023.7.10
timm==1.0.19
tokenizers==0.20.3
tomli==2.2.1
torch==2.0.1+cu118
torchmetrics==1.5.2
torchvision==0.15.2+cu118
tqdm==4.67.1
traitlets==5.14.3
transformers==4.46.3
trimesh==2.35.39
triton==2.0.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.2.3
wcwidth==0.2.13
Werkzeug==3.0.6
yapf==0.43.0
zipp==3.20.2
